requirements.txt
requests==2.32.3
python-dotenv==1.1.0
langchain-astradb==0.6.0
langchain-openai==0.3.12
langchain-community==0.3.20
langchain-core==0.3.50
fastapi[standard]==0.115.2
